+++ sycl/plugins/unified_runtime/pi2ur.cpp
@@ -10,32 +10,11 @@
 // TODO: remove when SYCL RT is changed to talk in UR directly
 
 #include <pi2ur.hpp>
-#include <ur.hpp>
 
-// Early exits on any error
-#define HANDLE_ERRORS(urCall)                                                  \
-  if (auto Result = urCall)                                                    \
-    return ur2piResult(Result);
-
+extern "C" {
 __SYCL_EXPORT pi_result piPlatformsGet(pi_uint32 num_entries,
                                        pi_platform *platforms,
                                        pi_uint32 *num_platforms) {
-
-  // https://spec.oneapi.io/unified-runtime/latest/core/api.html#zerplatformget
-
-  uint32_t Count = num_entries;
-  auto phPlatforms = reinterpret_cast<zer_platform_handle_t *>(platforms);
-  HANDLE_ERRORS(zerPlatformGet(&Count, phPlatforms));
-  if (*num_platforms) {
-    *num_platforms = Count;
-  }
-  return PI_SUCCESS;
+  return pi2ur::piPlatformsGet(num_entries, platforms, num_platforms);
 }
-
-__SYCL_EXPORT pi_result piPlatformGetInfo(pi_platform platform,
-                                          pi_platform_info param_name,
-                                          size_t param_value_size,
-                                          void *param_value,
-                                          size_t *param_value_size_ret) {
-  die("Unified Runtime: piPlatformGetInfo is not implemented");
-}
+} // extern "C
